Add files via upload
These are directly translated using Google Chrome to translate only the words from Chinese to English.  The function created errors in the syntax, deleting some parts and adding quotes which I edited each file line by line until they worked in SD.  I have translated all of the files from the original json folder as well the main so that most of the functions are usuable in english.  There were some buttons which I did not understand so they have been improvised.  This is "NOT" a perfect translation of the original and is my first attempt at helping those who asked for an english version.  I am finding it useful already, so I am forking it with only the 7 files which are working on my system currently along with the files needed to make the main buttons in english.  I also added a button for a negative prompt which is very popular with some additions to make it useful.  I hope this fork is useful to those in the SD world??!!